Script now respects platform

+++ install_emacs24.4.sh
@@ -1,17 +1,25 @@
 #!/bin/bash
-#taken from http://askubuntu.com/a/546050
 cd ~
-mkdir emacs-src && cd emacs-src
-wget http://mirror.team-cymru.org/gnu/emacs/emacs-24.4.tar.gz
-tar xvf emacs-24.4.tar.gz
 
-sudo apt-get install build-essential
-sudo apt-get build-dep emacs24
-cd emacs-24.4
-./configure
-make
-sudo make install
+#install emacs24.4
+platform=$(uname)
 
-rm -rf ~/emacs-src
+if [[ $platform == "Darwin" ]]; then
+    brew install emacs --with-cocoa --with-gnutls
+elif [[ $platform == "Linux"  ]]; then
+    taken from http://askubuntu.com/a/546050
+    mkdir emacs-src && cd emacs-src
+    wget http://mirror.team-cymru.org/gnu/emacs/emacs-24.4.tar.gz
+    tar xvf emacs-24.4.tar.gz
+
+     sudo apt-get install build-essential
+     sudo apt-get build-dep emacs24
+     cd emacs-24.4
+     ./configure
+     make
+     sudo make install
+     rm -rf ~/emacs-src
+fi
+
 cp ~/dotfiles/.emacs
 cp -r ~/dotfiles/.emacs.d/
